moving useFormatBalance out of the api module
