+++ src/ui/components/SubmitProposal.tsx
@@ -4,7 +4,7 @@ import {Button, Modal, Subheading, TextInput, Caption} from '@ui/library';
 import {Padder} from '@ui/components/Padder';
 import {SelectAccount} from '@ui/components/SelectAccount';
 import {useApiTx} from 'src/api/hooks/useApiTx';
-import {useFormatBalance} from 'src/api/hooks/useFormatBalance';
+import {useFormatBalance} from 'src/hooks/useFormatBalance';
 import globalStyles, {standardPadding} from '@ui/styles';
 import BalanceInput from './BalanceInput';
 import type {Account} from 'src/api/hooks/useAccount';
