+++ src/ui/screens/AddBountyScreen.tsx
@@ -14,7 +14,7 @@ import {useApi} from 'context/ChainApiContext';
 import {useSnackbar} from 'context/SnackbarContext';
 import {InputLabel} from '@ui/library/InputLabel';
 import {decimalKeypad} from 'src/utils';
-import {useFormatBalance} from 'src/api/hooks/useFormatBalance';
+import {useFormatBalance} from 'src/hooks/useFormatBalance';
 import {useBountiesSummary} from 'src/api/hooks/useBountiesSummary';
 import type {BN} from '@polkadot/util';
 import {formattedStringToBn} from 'src/api/utils/balance';
