+++ src/ui/screens/SendFundScreen.tsx
@@ -17,7 +17,7 @@ import {useChainInfo} from 'src/api/hooks/useChainInfo';
 import {InputLabel} from '@ui/library/InputLabel';
 import {useSnackbar} from 'context/SnackbarContext';
 import {BN_ZERO} from '@polkadot/util';
-import {useFormatBalance} from 'src/api/hooks/useFormatBalance';
+import {useFormatBalance} from 'src/hooks/useFormatBalance';
 import {stringToBn as stringToBnUtil, formattedStringToBn} from 'src/api/utils/balance';
 import AddressInput from '@ui/components/AddressInput';
